No longer using pagination
This is because there is no good way to split up pagination
for blog posts and for project posts. In the future with pagination,
pages of blog posts would be unusually short as project posts that
were included with that page would be hidden. Ideally I would split
posts into project versus blog posts and then paginate.